refactor: combine types/auth.ts and types.ts
