+++ frontend/src/contexts/AuthContext.ts
@@ -1,5 +1,5 @@
 import { createContext } from "react";
-import type { AuthContextType } from "../types/auth";
+import type { AuthContextType } from "../types";
 
 export const AuthContext = createContext<AuthContextType | undefined>(
   undefined,
