+++ frontend/src/hooks/useAuth.ts
@@ -1,6 +1,6 @@
 import { useContext } from "react";
 import { AuthContext } from "../contexts/AuthContext";
-import type { AuthContextType } from "../types/auth";
+import type { AuthContextType } from "../types";
 
 export function useAuth(): AuthContextType {
   const context = useContext(AuthContext);
