+++ frontend/src/types.ts
@@ -56,3 +56,18 @@ export interface FieldProps {
   label: string;
   value: string | null | undefined;
 }
+
+export interface UserData {
+  id: string;
+  name: string;
+  email: string;
+  uid?: string;
+}
+
+export interface AuthContextType {
+  user: UserData | null;
+  isLoading: boolean;
+  needsLogin: boolean;
+  logout: () => Promise<void>;
+  login: () => void;
+}
